requirements.txt
google-genai
Pillow
openai